add jni jnigraphics lib

+++ handler/src/main/java/com/ggg/handler/NativeLib.kt
@@ -5,5 +5,5 @@ object NativeLib {
         System.loadLibrary("handler")
     }
 
-    external fun changeImageGray(buff: ByteArray, width: Int, height: Int): ByteArray
+    external fun changeImageGray(buff: IntArray, width: Int, height: Int): IntArray
 }
